Kd tree (#14)
* tree squash to rebase

* Tree fix (#18)

Summary: kd_tree works, but works quite slow. Take a look in binary_search and in algorithm of building the tree.
Main problem: cant find good approach for working with rays that have their origs inside of any aabb. with them we should look through all nodes, taking needed figures and dists.
But in fact, when we compute something with ray that has his orig outside of a bounding box - everything looks fine.
* bug fix in plane cut and changed location of tree_trace

* start of fix in bin search

* still needs fix

* kd-tree works but too slow

* added shift to duck, changed logic in bin_search - returning pairs of needed figures and their dist_i. Picture has 4 ducks

* coube works, but bad reflections, and bad with kd-tree (#19)

* fiction - just changed params for picture

* rebase fixes